exchange connections munin plugin
tor_connections started blocking and I wasn't able to find the root
cause for it nor an updated version of the plugin.
This also blocked munin itself, which had the issue that the node
disappeared within munin.

Based on https://lists.torproject.org/pipermail/tor-talk/2006-June/010486.html
it seems to more or less match the open filedescriptors and hence
we monitor rather this than rely on a unmanageable plugin.

The only drawback is that this must run as root, as non-root users
can't read the filedescriptors from proc.