do not configure v3 and v2 onion services together
A onion service can't be v3 and v2 at the same time. At least
the hostname file would not be clearly attributed to a certain
version.

Alos since Tor version 0.3.5 the default version is v3 onion
services but only for new onion services. For existing services
it stays v2.

So let's either configure v2 or v3, but not both at the same time.